server: rename ill-named package types
and move to its proper location in server
